github PR comment - errors fix update
