+++ src/main/java/com/amogoscode/group/ebankingsuite/EBankingSuiteApplication.java
@@ -1,4 +1,4 @@
-package com.amogoscode.groupe.ebankingsuite;
+package com.amogoscode.group.ebankingsuite;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
