+++ src/main/java/com/amogoscode/group/ebankingsuite/exception/ResourceNotFoundException.java
@@ -1,4 +1,4 @@
-package com.amogoscode.groupe.ebankingsuite.exception;
+package com.amogoscode.group.ebankingsuite.exception;
 
 public class ResourceNotFoundException extends RuntimeException{
