+++ src/main/java/com/amogoscode/group/ebankingsuite/user/resquests/UserAuthenticationRequests.java
@@ -1,4 +1,4 @@
-package com.amogoscode.groupe.ebankingsuite.User.resquests;
+package com.amogoscode.group.ebankingsuite.user.resquests;
 
 public record UserAuthenticationRequests(
         String emailAddress,
