+++ src/test/java/com/amogoscode/group/ebankingsuite/EBankingSuiteApplicationTests.java
@@ -1,4 +1,4 @@
-package com.amogoscode.groupe.ebankingsuite;
+package com.amogoscode.group.ebankingsuite;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
