+++ src/test/resources/application.yml
@@ -1,7 +1,3 @@
-#server:
-#  port: 8080
-#  error:
-#    include-message: always
 
 spring:
   datasource:
